+++ fiboa_cli/__init__.py
@@ -14,7 +14,7 @@ from .rename_extension import rename_extension as rename_extension_
 from .validate import validate as validate_
 from .validate_schema import validate_schema as validate_schema_
 from .version import __version__, fiboa_version as fiboa_version_
-from .util import log, check_ext_schema_for_cli, valid_file_for_cli, valid_file_for_cli_with_ext, valid_files_folders_for_cli, valid_folder_for_cli
+from .util import log, check_ext_schema_for_cli, parse_converter_input_files, valid_file_for_cli, valid_file_for_cli_with_ext, valid_files_folders_for_cli, valid_folder_for_cli
 
 @click.group()
 @click.version_option(version=__version__)
@@ -334,6 +334,14 @@ def jsonschema(schema, out, fiboa_version, id_):
     required=True
 )
 @click.option(
+    '--input', '-i',
+    type=click.STRING,
+    help='File(s) or URL(s) to read from. Can be used multiple times. Specific files from ZIP and 7Z archives can be picked by providing the archive path and the file path in the archive separated by an equal sign. To pick multiple files from a single archive separate them by comma. Example: /path/to/archive.zip=file1.gpkg,subfolder/file2.shp',
+    callback=parse_converter_input_files,
+    multiple=True,
+    default=None
+)
+@click.option(
     '--cache', '-c',
     type=click.Path(exists=False),
     help='By default the CLI downloads the source data on every execution. Specify a local folder to avoid downloading the files again. If the files exist, reads from there, otherwise stores the files there.',
@@ -358,13 +366,13 @@ def jsonschema(schema, out, fiboa_version, id_):
     help='Compression method for the Parquet file. Defaults to zstd.',
     default="zstd"
 )
-def convert(dataset, out, cache, source_coop, collection, compression):
+def convert(dataset, out, input, cache, source_coop, collection, compression):
     """
     Converts existing field boundary datasets to fiboa.
     """
     log(f"fiboa CLI {__version__} - Convert '{dataset}'\n", "success")
     try:
-        convert_(dataset, out, cache, source_coop, collection, compression)
+        convert_(dataset, out, input, cache, source_coop, collection, compression)
     except Exception as e:
         log(e, "error")
         sys.exit(1)