+++ fiboa_cli/datasets/de_sh.py
@@ -26,13 +26,14 @@ MISSING_SCHEMAS = {
     }
 }
 
-def convert(output_file, cache = None, source_coop_url = None, collection = False, compression = None):
+def convert(output_file, input_files = None, cache = None, source_coop_url = None, collection = False, compression = None):
     """
     Converts the Schleswig-Holstein (Germany) field boundary datasets to fiboa.
     """
     convert_(
         output_file, cache,
         SOURCES, COLUMNS, ID, TITLE, DESCRIPTION,
+        input_files=input_files,
         extensions=EXTENSIONS,
         missing_schemas=MISSING_SCHEMAS,
         license=LICENSE,